Using Portal for the Fullscreen mode
