+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ssa-ui-kit/core",
-  "version": "1.1.15",
+  "version": "1.1.16",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "private": false,
