+++ packages/core/src/components/Charts/BarLineComplexChart/BarLineComplexChartTooltip.tsx
@@ -51,6 +51,7 @@ export const BarLineComplexChartTooltip = forwardRef<
         display: 'flex',
         flexDirection: 'column',
         alignItems: 'flex-start',
+        zIndex: 1,
       }}>
       {data.map((item) => {
         const color = path(['marker', 'color'])(item) as string | undefined;
